Security - only admins can delete - check for sessions on sensitive actions - insert server-side checks for values
finishes
- Correction bug d'affichage divers (titres, maj, ortho...)
- Boutons supprimé que pour les admin
- boutons annuler partout

+++ dal/balises.js
@@ -3,6 +3,7 @@ var express = require('express');
 var router  = express.Router();
 
 router.post('/create', function(req, res) {
+
     models.Balise.create({
         nom: req.body.nom,
         indice: req.body.indice,
@@ -13,12 +14,62 @@ router.post('/create', function(req, res) {
     });
 });
 
+
 router.get('/destroy/:Balise_id', function(req, res) {
-    models.Balise.destroy({
-        where: { id: req.params.Balise_id }
-    }).then(function() {
-        res.redirect('/Balises/view');
-    });
+
+    if (!req.session.admin) {
+        res.render('error', {
+            message: "Accès refusé",
+            error: "Accès refusé"
+        });
+
+    } else {
+
+    models.Ptobq.findAll({
+        where: { BaliseId: req.params.Balise_id }
+    }).then(
+        function(ents) {
+
+            if (ents.length > 0) {
+
+              models.Balise.findOne({
+                  where: { id: req.params.Balise_id }
+              }).then(
+                  function(balise) {
+                      res.render('balises_detail', {
+                          balise: balise,
+                          err: "Vous ne pouvez pas supprimer cette balise car elle est associée à des parcours"
+                      });
+                  });
+
+            } else {
+
+                models.Balise.destroy({
+                    where: { id: req.params.Balise_id }
+                }).then(function() {
+
+                    models.Balise.destroy({
+                        where: { id: req.params.Balise_id }
+                    }).then(function() {
+
+                        models.Balise.findAll().then(
+                            function(balises) {
+                                res.render('balises_view', {
+                                    balises: balises,
+                                    ok: "La balise à correctement été supprimée"
+                                });
+                            });
+
+                    });
+
+                });
+
+            }
+        });
+
+   }
+
+
 });
 
 router.post('/update/:Balise_id', function(req, res) {
